Allow overriding the command line when pinning to the taskbar
Windows 7 and later allow applications to be pinned to the taskbar. In
particular when calling mintty as the work horse of Git for Windows'
"Git Bash", we will want to re-launch the Git Bash when being pinned, not
just mintty.

Let's just add a setting to allow users to override which command gets
to be launched when the pinned mintty is called, and also a setting to
configure the display name in the task bar.

Note: the RelaunchCommand setting will be ignored unless the AppID *and*
RelaunchDisplayName are configured, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>