Set the app ID associated with the window explicitly
When calling mintty from another process and painstakingly configuring
the app ID to be identical between the caller and the callee, at least
on this developer's machine clicking the icon pinned to the taskbar
*still* launched mintty directly, not the caller.

However, if we set the app ID explicitly in the window's property store
(in addition to setting the current process' app ID), clicking a pinned
icon of an executable that called mintty will actually launch that
executable, as obviously intended. So let's set the app ID of the mintty
window explicitly.

We now need to declare the minimum Windows version to be Windows 7 so we
get the necessary data structures, but we take pains to make sure that
`mintty` will still run on previous Windows versions by repeating the
`LoadLibrary()`/`GetProcAddress()` dance that we already perform when
calling `SetCurrentProcessExplicitAppUserModelID()`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>