refactor: unify attributes-to-colours logic (no-op)
Logic which turn attributes into display colours were duplicated in
several places. For instance, five places did bold-to-colour, three
places did dim-to-colour, and more.

This maade it hard to modify such procedures, as the changes would
have to be duplicated, or else we get inconsistent behavior.

This commit puts most of the attribute-to-colour logic at one function
which _can_ do all of them, but does different subsets of them depending
on a mode. Each original place which duplicated such logic has a mode of
its own which reflects its original process and the attributes it applied.

Hopefully the number of modes can be reduced over time.

The behaviour should remain the same, and hopefully make it easier to
modify attribute-to-colours logic, as it's all in one place now.