bugfix: avoid read/write out-of-bounds when copying true colour as rtf
The copy-as-rtf code completely ignores the existance of true colours. It
doesn't even have their values, as it works on an array of attributes rather
than of cattr items - so it doesn't have access to truefg and truebg .

It assumes fgi/bgi are always smaller than COLOUR_NUM and uses them as indices
while writing/reading from the "palette" array, which has COLOUR_NUM items.

But when fgi/bgi are true colour, it's an out-of-bounds access
(the TRUE_COLOUR fgi/bgi value is bigger than COLOUR_NUM).

This issue existed also prior to the refactor at the previous commit.

This commit makes sure that it uses default fg/bg colours on such cases such
that it always accesses valid palette indices, but a proper fix would be
to handle true colours rather than replacing them with an unrelated values.