dim: simplify, improve, decouple from BoldAsColour
TL;DR: This commit changes dim to do only this:
- Blend the cell's foreground with the terminal's default background
  (term.rvideo is considered for the default bg).

The long version:

It seems impossible to track the reasons which led to the dim behavior
before this commit, but one might still try, starting from 2009:
- cc7e198d Dim support introduced - only half absolute intensity for fg.
- ce8ac70a if bold-as-font is set, blend with the cell's bg (why?)
- aa821e07 bold ANSI is coloured, dim-blend only default colour (what?!?)
- 964b3097 Support both BoldAsFont and AsColour. Change dim (why?)
- (later)  True colours added. Dim's >= 256 now includes them (accident?)

The first two commits are no longer at this repository, as they were
merged from branches which were since deleted. They do exist at other
repositories, and they appear at this repo as parts of merges 31b36bc0 and
6ce15c43 respectively, but without the original commit messages.

Not even one of those commits (including the original commit messages
before the merges) gives any kind of explanation on WHY the change was made.

Which brings us to today before this commit:

If BoldAsColour=no and/or it's the default colour and/or it's a true colour:
- blend the fg with the cell's bg.
Otherwise:
- Halve the absolute intensity of the fg colour.

Regardless of reasons, it got few, well.. all things wrong:
- There should be no condition which depends on the colour class (default or
  true colour etc). Dim is a brightness thing which depends on RGB values.

- Blend with the cell background only sort of works with non coloured
  background (gray/black/white). With coloured background the fg colour
  changes in an unexpected way for "dim".

- Half absolute brightness is incorrect: too dark on gray-ish background,
  plain wrong with bright background (makes it actually look more intense).

This commit, other than making it consistent and better, also finally
separates dim from bold after a decade of being tangled for no good reason.

In the future we might want to consider:
- Also blend the cell's background with the terminal's default background
  to get a more even dimmed look, and maybe closer to what people expect.

General note:
- If the code just doesn't make sense, it could be a divine message
  channeled through past developers, but it could also simply be a bug.