remove truecolour(..), use attr_apply_colour(..) directly
Now that dim doesn't use a custom "against" as it works against the
default terminal background, truecolour(..) doesn't do anything unique
anymore. Its only unique feature was supporting a custom bg which was
only used for dim.

Therefore the two clients which used it can now use attr_apply_colour
directly without having to provide a custom background (it was effectively
never custom, with a remote possiblity to be custom only for one of them).