bold: update docs
It's inherently complex to explain how the bold attribute is displayed
and configured, but for this exact reason we owe our users a good
description.

Let's split it such that the BoldAsFont and BoldAsColour sections focus
on what bold font or different colour mean, while adding another section
which explains how they affect the different colour classes.