Implement basic per-window v2 DPI awareness
This enables the user to drag mintty from one screen to another screen with
a different pixel density without any fluttering.

Unfortunately, the size of the terminal (rows, cols) still changes when the
pixel density changes. The suggested way to fix this properly is to handle
WM_GETDPISCALEDSIZE and return the desired window dimensions. This means that
we have to compute the new window dimensions in terms of the old dimensions,
which requires a way to transform dim -- old dpi -> (rows, cols) -- new dpi
--> dim', including the old and new size of the search and scroll bar.

But this change already makes the options dialog pop up at the right size.
Previously, it was always unscaled on my HiDPI screen.