loop-drain terminal input for less flickering and work
Previously we were trying to read once up to 4K input bytes, but
underlaying buffering could limit us. E.g. MSYS2 gives up to 256 bytes.

If an application outputs a bigger chunk, then it can result in partial
terminal/screen updates which typically manifest as flickering or
tearing or "not solid" screen updates etc.

This usually also meant more work because we will also update the screen
again once we read all the input.

Change it to drain the input in a loop as long as we have input, up to
the same 4K bytes as before.

This results is much less flickering of the status line in 'less' (and
therefore 'man' etc), and generally a more "solid" feeling on updates.