Avoid nested functions
It was reported to the oss-security mailing list in
https://seclists.org/oss-sec/2018/q4/82 that GCC 7 and 8 have known
issues generating code for nested functions, essentially voiding any
Data Execution Protection (DEP).

As it is unclear how quickly GCC can be fixed, let's just avoid the
nested functions altogether.

Note: this patch mainly moves the nested functions out of their previous
host function, if necessary adds a suffix to the function name, and in
several cases declares structures to be passed via the `lParam` argument
of the `Enum*()` functions, to retain access to the local variables of
the former host function.

Note, too, that trailing whitespace was preserved faithfully when moving
and de-indenting the code.

Also please note that this patch does *not* care about code that is not
compiled because it is hidden behind `#if ... #endif` guards.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>