tabbar: Fix startup position (#1044)
When tabbar is enabled, users may expect that all mintty windows are
placed at the same position.
If Alt+F2 is used, the new window is placed at the same position as the
current window. However, if mintty is executed in another way, the
position is not the same as the current window.

This fixes the position of a new window even if Alt+F2 is not used.