tabbar: Don't recreate the tabbar every time (#1043)
The height calculation in win_toggle_tabbar() was wrong. So, when the
function was called, the tabbar was recreated every time,
by the condition `if (height != TABBAR_HEIGHT && initialized)`.

Because of this, changing the title sometimes took very long time.
(With the default installation of Cygwin, the current directory is shown
as the title. Thus, on my slow machine, executing a `cd` command
sometimes took more than one second.)

Fix the height calculation and remove an unused variable.