Separate foreground and background values for ANSI colours (#1151)
Add the ability to specify separate foreground and background values for
the 16 ANSI colours, so that e.g. blue text can be bright enough to be
readable on a dark background, while background blue can be dark enough
to not interfere with light foreground text, e.g. in Midnight Commander.

The values can be given as semicolon-separated pairs in config files.
When only one value is present, it is used for both the foreground and
background, to ensure backward capability with existing themes.

Also add mintty-specific control sequence OSC 7765 for setting and
querying the foreground and background variants of the ANSI colour
values. This sequence does not touch the first sixteen xterm256 colours
that can be controlled through the OSC 4 and 104 sequences. Conversely,
though, for backward compatibility those xterm sequences continue to
control the ANSI colours as well, setting the foreground and background
variants to the same value.

Put the new feature to use in the default 'helmholtz' theme, darkening
background blue a lot and the other non-bold colours by a bit as well,
to improve readability of light text on them. Except for yellow, because
that isn't suitable as background for light text anyway.

Repurpose the 'kohlrausch' theme as a black-on-white one, with darker
foreground than background colour values.

Reflect changes in documentation and bump copyrights.