Add Alt-modified mousewheel mode using DECSET 7765
When enabled, this sends Alt-modified cursor up/down keycodes for
mousewheel events. If the Alt key is pressed, unmodified keycodes are
sent. The mode does not affect page up/down codes.

In the nano editor from version 4.0, the Alt+Up/Down keycodes scroll the
window instead of moving the cursor. Scrolling straight away is the
proper action for the mousewheel, whereas moving the cursor only scrolls
once the cursor reaches the top or bottom.

Less can be configured to scroll on Alt+Up/Down with this in ~/.lesskey:
\e[1;3A back-line
\e[1;3B forw-line

Similarly, vim can be configured to do so with this in ~/.vimrc:
map <M-Up> <C-Y>
map <M-Down> <C-E>
imap <M-Up> <C-O><C-Y>
imap <M-Down> <C-O><C-E>