Support lone Alt key presses in KeyFunctions
A lone Alt key press in Windows applications normally brings up the
application menu bar (which is why its virtual keycode is VK_MENU).
There's no such menu bar in mintty, which relies on the context menu
and window menu instead, so make lone Alt key presses available for
mapping in the KeyFunctions setting.

The action is triggered when the Alt key is released, but only if no
other modifier was down when the Alt key was pressed, and if nothing
else was pressed in-between.

Also make sure that AltGr does not trigger this action or Alt code
processing, unless the AltGrIsAlsoAlt setting is enabled.

One use case for this is to map Alt to the context menu via the
'menu-text' action, to make it easier to reach on keyboards that don't
have the context menu key.

Another is to map it to ^[, i.e. the Esc key code, which might be
convenient for vim users. It also makes plenty of sense as Alt modifier
combinations normally prefix the modified key with ^[.

As the Esc key's default function can be replaced in this way, also make
unmodified Esc key presses available for mapping in KeyFunctions. One
use case for that is to map it to a non-letter control character such as
^\, and then use that as the interrupt character in the stty settings,
instead of ^C. That gives Esc its usual Windows meaning of interrupting
or cancelling actions, and makes ^C available for copying.